+++ App.js
@@ -2,12 +2,10 @@
 import "react-native-gesture-handler";
 import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
 import { Provider as PaperProvider ,DefaultTheme } from "react-native-paper";
-// screens
-import Home from "./screens/Home.js";
 
-import Bn from "./components/BottomNavigation";
+
+import BottomNavigation from "./navigation/BottomNavigation";
 
 const theme = {
   ...DefaultTheme,
@@ -21,7 +19,7 @@ export default function App() {
   return (
     <PaperProvider theme={theme}>
       <NavigationContainer>
-        <Bn />
+        <BottomNavigation />
       </NavigationContainer>
     </PaperProvider>
   );
