--- components/BottomNavigation.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-import { createMaterialBottomTabNavigator } from "@react-navigation/material-bottom-tabs";
-
-const Tab = createMaterialBottomTabNavigator();
-
-import Home from "../screens/Home";
-import SettingsScreen from "../screens/StudentDashboard";
-
-export default function MyTabs() {
-  return (
-    <Tab.Navigator>
-      <Tab.Screen name="Home" component={Home} />
-      <Tab.Screen name="Settings" component={SettingsScreen} />
-    </Tab.Navigator>
-  );
-}
