+++ screens/Login/Login.js
@@ -1,27 +1,22 @@
 import React from "react";
 import { View, Text, FlatList } from "react-native";
 // import { Colors } from "react-native/Libraries/NewAppScreen";
-import Box from "./PreviewBox";
-const Preview = ({ name, color }) => {
+import ColorBox from "../../navigation/ColorBox";
+
+const Login = ({ route }) => {
+  const { name, COLORS } = route.params;
   return (
     <View>
-      <Text>{name}</Text>
+      <Text>This is the Login page : {name}</Text>
       <FlatList
-        data={color.slice(0,4)}
-        renderItem={(data) => <Box color={data.item.hexCode} />}
-        keyExtractor={data=>data}
-        numColumns={5}
-        
-      />
-      {/* <FlatList
         data={COLORS}
         renderItem={(data) => (
           <ColorBox colName={data.item.colorName} colCode={data.item.hexCode} />
         )}
         keyExtractor={(item) => item}
-      /> */}
+      />
     </View>
   );
 };
 
-export default Preview;
+export default Login;
